Finish migration to BudgetAllowance and doing math in cents
In hindsight I probably should have broken this down into multiple diffs but at this point
it's all too intermingled to bother. Next time.

+++ src/routes/api/expense.rs
@@ -129,7 +129,7 @@ pub async fn delete_expenses(
 
 fn to_simple_csv_row(
     transaction_date: &str,
-    amount: f64,
+    amount: i32,
     description: &str,
 ) -> anyhow::Result<String> {
     let mut wtr = csv::WriterBuilder::new()
